clm4_5_11_r189
Instead of reading in maximum building temperature from surface dataset, read it in from
a time-varying urban streams file. This allows the urban Air Conditioning use to change
in time as well as spatially. Other spatially and temporally changing parameters can be added
(or moved from the surface dataset) to this stream file as well in the future. The new streams
file has the same kind of controls as other streams file for years to run over, filename, and
mapping algorithm. The time interpolation is assumed to be linear, and it uses taxmode=extend
so it will use the first value before the first year and the last value after the last year.
The current dataset is at one degree resolution (f09_g16). Because of this CLM45 compsets
will change answers to some extent (because interpolation methods for streams are slightly
different) -- but answers will be very similar. CLM50 compsets will change because there is
variation in the max building temperature both spatially and temporally.

Also fix a problem with the humanstress module for GSWP3 forcing because of temperatures that
are too high for the WetBulb method to converge. WetBulb previously only allowed two iterations
and now it allows more (10,000). If the method fails to converge it will exit with a warning.
This was worked on by Jonathan Buzan with Keith Oleson.