clm4_5_13_r204
This brings in some simple answer changing bug fixes. And some non-answer changing bug fixes for mksurfdata.pl and
the build-namelist unit tester.

In order to calculate infiltration the fraction of ground covered in surface water is used, before it is updated
to account for snow. Also mulitply k_wet by a factor of 1x10^-4 both of these will makes a difference in surface runoff.
This change is essentially a bug in the original formulation and so we've applied to both CLM45 and CLM50 answers.
Since this is hydrology it will affect answers in all CLM45/CLM%0 cases whether SP, CN or BGC modes.

For MEGAN we were using a new updated dataset that had the addition of all 78 PFT's including the new crop ones.
Unfortunately it also had bare-soil, and this moved the indexing off by one. So we have a new dataset that removes
the bare-soil and hence works correctly now.

A simple addition of setting lake energy flux (eflx_grnd_lake) for the case when there isn't one layer os snow.

For CLM45 we found that the use of the Zheng rooting profile for Carbon resulted in 1850 spinup simulations that couldn't
spinup the tropics. Using the Jackson-1996 profile alleviates this problem.

There were also some simple problems in using the queryNamelistDefaults.pl tool with the new cime tag brought in
with clm4_5_12_r202. Fixing this allows mksurfdata.pl and build-namelist-unit_tester to work.