clm4_5_16_r259
Update CLM to use version 3 for config_components.xml which has additional error checking of compset names. Also
update it for mosart, cism and rtm. And cime is updated to the version used in cesm2_0_beta07. There are a few
new features brought in with the latest cime that are described below. CLM now requires one and only one modifier
for what type it is in the compset name (i.e. only one "%" as in CLM45%BGC or CLM50%SP). Will now abort when
CLM build-namelist encounters a warning (after printing information on all warnings). To ignore the warnings
and continue add the new "-ignore_warnings" to CLM_BLDNML_OPTS. "-irrig" is now an option to CLM_BLDNML_OPTS
only for clm4_0, set "irrigate" namelist item in your user_nl_clm file for clm4_5 or clm5_0. By default build-namelist
sets irrigate to true for transient or present-day configurations if crop is on and clm5_0 (unless DV on). Get PTCLMmkdata fully
converted over to cheyenne, and create a new set of files for testing it's ability to create cases. Checking for
.true. or .false. in CLM build-namelist is now more robust and allows for "F" or "T" as newer FORTRAN versions do.

Fix several bugs as described below. PTCLM and CLM tools were both verified to work and pass testing.
The PE layout for f10 on hobart was changed to a single-node so that it would compare exactly to previous versions.
Fix a bug for NCK tests in RTM and MOSART. Get mksurfdata_map working correctly for f05. Error message when
you set finidat=' ' is more clear on what to do. You can also now do a cold-start for a hybrid case prints a warning),
but not allowed on a branch case.