Allow holes in the number of history tapes
Holes are cases where, for example, we have h0, h1 and h3 tapes, but no
h2 tape (because there are no fields on the h2 tape).

The key part of this implementation is cycling some loops if a given
history tape is not in use. There are a few other loops for which it
probably would have been fine to do this cycling, but it wasn't
necessary - e.g., doing initialization of some per-tape variables. In
most of those cases, I didn't bother doing this cycling because it
seemed like it could keep the code more robust to still initialize these
per-tape variables for the not-in-use tapes.

Note that I removed this error check:

    ! Check that the number of history files declared does not exceed
    ! the maximum allowed.

    if (ntapes > max_tapes) then
       write(iulog,*) trim(subname),' ERROR: Too many history files declared, max_tapes=',max_tapes
       call endrun(msg=errMsg(sourcefile, __LINE__))
    end if

because, with the way ntapes is determined, it's impossible for ntapes
to exceed max_tapes - i.e., this error check wasn't actually doing
anything.

Also note that I put in some error-checking code to ensure that, in a
restart run, if a history file is in use in the current run, then it was
in use on the restart file, and vice versa.