Don't allow interpolation from a non-ciso case to a ciso case
Due to https://github.com/ESCOMP/ctsm/issues/67, interpolation from a
case without carbon isotopes to a case with carbon isotopes yields
incorrect initialization values for the carbon isotopes. Now that we're
turning carbon isotopes on via some semi-out-of-the-box usermods (for
cmip6), it is becoming more important to check to make sure someone
doesn't shoot themselves in the foot this way.

This commit should be backed out once ESCOMP/ctsm#67 is resolved