Add soilm_ignore_data_if_missing
Add new namelist item soilm_ignore_data_if_missing and react to it.
Also use old way for the indexing in an attempt to reproduce the
previous results. If soilm_ignore_data_if_missing is true and find
a missing point expected continue calculation as before. If not
abort when that point is found. Also comment out the section
where values greater than 10 were set to spval as that changes
answers.