Checkin version that is identical to release-clm5.0.27
OK, this is a version that is identical to release-clm5.0.27. Found out
the issue was that lai_interp can't be moved the way I did. It needs
to be divided into the streams part and non-streams part like is
being done in SoilMoistureStreams.

So this replicates answers for SMS_P720x1_D_Ld3.f09_g17.I2000Clm50SpGs.cheyenne_intel.clm-prescribed